fix: move fixture loading to root conftest.py
Move django_db_setup fixture that loads character fixtures to the root
conftest.py to ensure fixtures are discovered and loaded properly across
all test workers in pytest-xdist parallel execution.
